refactor : domain보다는 View에서 입력받는 내용이라 판단하여 package 이동
