+++ src/main/java/view/ContinuePlay.java
@@ -1,4 +1,4 @@
-package domain;
+package view;
 
 public class ContinuePlay {
